refactor: moved all components in component folder to domain specific folder
