+++ src/domains/product/components/productBoard/index.tsx
@@ -3,12 +3,11 @@
 import Link from "next/link";
 import { useState } from "react";
 
+import AddToCartButton from "@/domains/store/shoppingCard/components/addToCartButton";
+import Quantity from "@/domains/store/shoppingCard/components/quantity";
 import { StarIcon, HeartIcon } from "@/shared/components/icons/svgIcons";
 import { TProductBoard } from "@/types/product";
 import { TCartItem } from "@/types/shoppingCart";
-
-import Quantity from "../../common/quantity";
-import AddToCartButton from "../addToCartButton";
 
 const ProductBoard = ({ boardData }: { boardData: TProductBoard }) => {
   const { name, id, isAvailable, specialFeatures, price, shortDesc, dealPrice, defaultQuantity } = boardData;
