+++ src/domains/product/components/productCard/index.tsx
