+++ src/domains/store/shared/components/footer/index.tsx
