+++ src/domains/store/shared/components/navbar/navFavorite/index.tsx
