+++ src/domains/store/shared/components/warning/index.tsx
