+++ src/domains/store/shoppingCard/components/shoppingCart/_components/cartItem/index.tsx
